+++ test/express/src/schemas/request_log_schema.js
@@ -1,0 +1,16 @@
+const {Schema} = require('mongoose');
+
+const requestLogSchema =  new Schema({
+    req_time: Date
+},{ 
+    autoIndex: false,
+    timestamps: {
+        createdAt: 'created_at',
+        updatedAt : 'update_at'
+    },
+    strict: false
+});
+
+
+
+module.exports = requestLogSchema;